# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/stores/{storeId}/notifications": {
      "get": {
        "operationId": "NotificationsController_findAll",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "read",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "archived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/stores/{storeId}/notifications/unread-count": {
      "get": {
        "operationId": "NotificationsController_unreadCount",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationCountResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/stores/{storeId}/notifications/{notificationId}/read": {
      "patch": {
        "operationId": "NotificationsController_markRead",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "notificationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/stores/{storeId}/notifications/read-all": {
      "post": {
        "operationId": "NotificationsController_markAllRead",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationCountResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/stores/{storeId}/notifications/{notificationId}/archive": {
      "patch": {
        "operationId": "NotificationsController_archive",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "notificationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/contact": {
      "post": {
        "operationId": "ContactController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInquiryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InquiryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ]
      },
      "get": {
        "operationId": "ContactController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InquiryResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ]
      }
    },
    "/contact/{id}/review": {
      "patch": {
        "operationId": "ContactController_markReviewed",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InquiryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bias Market API",
    "description": "Bias Market — niche-first store builder API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "NotificationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "type": {
            "enum": [
              "LOW_STOCK",
              "OUT_OF_STOCK"
            ],
            "type": "string"
          },
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "read": {
            "type": "boolean"
          },
          "readAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "archived": {
            "type": "boolean"
          },
          "archivedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "storeId",
          "type",
          "entityType",
          "entityId",
          "title",
          "body",
          "metadata",
          "read",
          "readAt",
          "archived",
          "archivedAt",
          "createdAt"
        ]
      },
      "NotificationCountResponseDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number"
          }
        },
        "required": [
          "count"
        ]
      },
...
      "CreateInquiryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "company": {
            "type": "string"
          },
          "inquiryType": {
            "type": "string"
          },
          "message": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "name",
          "email",
          "message"
        ]
      },
      "InquiryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "company": {
            "type": "string",
            "nullable": true
          },
          "inquiryType": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string"
          },
          "status": {
            "enum": [
              "NEW",
              "REVIEWED",
              "ARCHIVED"
            ],
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "company",
          "inquiryType",
          "message",
          "status",
          "createdAt"
        ]
      },
...
    }
  }
}